Create Winium for AE citrix
This is a simple unit test project.
We need winium dll files: 
Winium.Elements.Desktop
Winium.WebDriver.
If required we can add:  UIAutomationTypes

Process.start : this will start the winium instance. we need to download and keep the winium at the desired location.
setCapability : will set the app path. here the app is AE120.ica.
RemoteWebDriver: This will open the instance of application.